Updated dependencies, added integrated logger system, change default password to generated on first run
