+++ .env.prod
@@ -1,5 +1,6 @@
 PYTHONPATH=./home_warehouse_api
 DEBUG=False
+TEST=False
 DB_URL=mongodb://127.0.0.1:27017/home-warehouse
 API_HOST=127.0.0.1
 API_PORT=8000
