+++ .python-version
@@ -1,1 +1,1 @@
-3.9.0
+3.9.1
