+++ README.rst
@@ -9,13 +9,15 @@ With `docker <https://docs.docker.com/engine/install/>`_ (recommended)
 
   #. Run from terminal: ``bash <(curl -s https://raw.githubusercontent.com/Home-warehouse/warehouse-api/master/install_nix.sh) './hw' 'v0.3.2-alpha' 'localhost' 5001 5000``
   #. Go to http://localhost:5000/
-  #. Login to admin account, email: ``home-warehouse@mail.com``; password: ``home-warehouse-supervisor`` and change account data
+  #. See home-warehouse logs to copy generated password
+  #. Login to admin account with username: ``home-warehouse@mail.com`` and generated password, after it, it is advised to change account password
 
 - Windows
 
   #. Run from PowerShell: ``Invoke-WebRequest https://raw.githubusercontent.com/Home-warehouse/warehouse-api/master/install_windows.ps1 -OutFile .\install_windows.ps1; .\install_windows.ps1 './hw' 'v0.3.2-alpha' 'localhost' 5001 5000``
   #. Go to http://localhost:5000/
-  #. Login to admin account, email: ``home-warehouse@mail.com``; password: ``home-warehouse-supervisor`` and change account data
+  #. See home-warehouse logs to copy generated password
+  #. Login to admin account with username: ``home-warehouse@mail.com`` and generated password, after it, it is advised to change account password
 
 **Use evernote integration**
 
@@ -39,7 +41,8 @@ Standalone (Advanced)
 #. Setup nginx server (example config is in ``home-warehouse-ui/nginx``)
 #. Move files from ``dist`` directory to nginx hosted files directory
 #. Go to http://localhost:5000/
-#. Login to admin account, email: ``home-warehouse@mail.com``; password: ``home-warehouse-supervisor`` and change account data
+#. See home-warehouse logs to copy generated password
+#. Login to admin account with username: ``home-warehouse@mail.com`` and generated password, after it, it is advised to change account password
 
 
 ===========
